+++ project/app/src/main/java/org/owntracks/android/ui/status/logs/LogPalette.kt
@@ -3,4 +3,10 @@ package org.owntracks.android.ui.status.logs
 /**
  * Colour palette for log lines
  */
-data class LogPalette(val default: Int, val debug: Int, val info: Int, val warning: Int, val error: Int)
+data class LogPalette(
+    val default: Int,
+    val debug: Int,
+    val info: Int,
+    val warning: Int,
+    val error: Int,
+)